refactor: move Postal::QueryString to app/util/query_string
